Just refactoring
